"Use selector to get current user and render Outlet if user is present, otherwise render null"
Explanation:
The updated code uses the `useSelector` hook from `react-redux` to get the `currentUser` from the `user` state of the Redux store. If the `currentUser` exists, it renders the `Outlet` component from `react-router-dom`. Otherwise, it renders `null`. This is a change from the previous version, which always rendered a `div` with the text "PrivateRouter".