"Update PrivateRouter to redirect to signin on unauthorized access"
This commit updates the PrivateRouter component to redirect to the signin page when a user tries to access a protected route and is not currently logged in. Previously, it would simply return null, rendering nothing. Now, it uses the Navigate component from react-router-dom to redirect the user to the signin page.