"Update dependency of location in useEffect"
Explanation:
The code diff shows that the second argument of the useEffect hook in the Dashboard.jsx file has been updated from [location] to [location.search]. This change means that the useEffect hook will now only re-run if the search property of the location object changes, instead of re-running whenever the entire location object changes. This can help optimize the performance of the component by reducing unnecessary re-renders.