"Add state for current tab and use location search parameter to set it"
This commit updates the DashSidebar component to include state for the current tab and uses the search parameter of the location object from react-router-dom to set the tab when the component is rendered. This allows for the tab to be set based on the URL, so that when a user navigates to a specific tab, that tab will be selected when the page loads.