"Changed profile picture div to be cursor-pointer"
This commit message describes the change in the code diff for the `DashProfile` component where the `div` containing the user's profile picture has been modified to have the CSS property `cursor-pointer`. This suggests that the profile picture may now have some interactivity associated with it, such as being clickable or hoverable. However, the specifics of this change are not clear from the diff alone and would require additional context.